pyproject.toml
[project]
name = "yankee-stadium-beer-controls"
version = "0.1.0"
description = "Economic analysis of beer price controls at Yankee Stadium"
authors = [
    {name = "Max Ghenis", email = "max@policyengine.org"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["economics", "sports", "policy", "price-controls", "externalities"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "build>=1.2.2",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-pyyaml>=6.0.12",
]

[project.urls]
Homepage = "https://github.com/maxghenis/yankee-stadium-beer-price-controls"
Repository = "https://github.com/maxghenis/yankee-stadium-beer-price-controls"
Documentation = "https://github.com/maxghenis/yankee-stadium-beer-price-controls/blob/main/README.md"

[project.scripts]
yankee-beer-sim = "yankee_stadium_beer_controls.example:main"
yankee-beer-paper = "yankee_stadium_beer_controls.paper:main"
yankee-beer-calibrate = "yankee_stadium_beer_controls.calibrate:main"
yankee-beer-web-data = "yankee_stadium_beer_controls.web:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yankee_stadium_beer_controls"]

[tool.hatch.build.targets.wheel.force-include]
"config.yaml" = "yankee_stadium_beer_controls/default_config.yaml"
"paper/_quarto.yml" = "yankee_stadium_beer_controls/paper_project/_quarto.yml"
"paper/index.qmd" = "yankee_stadium_beer_controls/paper_project/index.qmd"
"paper/references.bib" = "yankee_stadium_beer_controls/paper_project/references.bib"

[tool.hatch.build.targets.sdist]
include = [
    "src/yankee_stadium_beer_controls",
    "config.yaml",
    "paper/_quarto.yml",
    "paper/index.qmd",
    "paper/references.bib",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--verbose",
    "--cov=yankee_stadium_beer_controls",
    "--cov-report=term-missing",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["yankee_stadium_beer_controls"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
